refactor(api): move random util to src/shared
