+++ api/src/certification/flash-certification/application/scenario-simulator-controller.js
@@ -5,9 +5,9 @@ import _ from 'lodash';
 import { HttpErrors } from '../../../../lib/application/http-errors.js';
 import { pickAnswerStatusService } from '../../../../lib/domain/services/pick-answer-status-service.js';
 import { scenarioSimulatorBatchSerializer } from '../../../../lib/infrastructure/serializers/jsonapi/scenario-simulator-batch-serializer.js';
-import { random } from '../../../../lib/infrastructure/utils/random.js';
 import { parseCsv } from '../../../../scripts/helpers/csvHelpers.js';
 import { pickChallengeService } from '../../../evaluation/domain/services/pick-challenge-service.js';
+import { random } from '../../../shared/infrastructure/utils/random.js';
 import { extractLocaleFromRequest } from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { FlashAssessmentSuccessRateHandler } from '../domain/models/FlashAssessmentSuccessRateHandler.js';
 import { usecases } from '../domain/usecases/index.js';
