+++ api/tests/shared/unit/infrastructure/utils/random_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import { random } from '../../../../lib/infrastructure/utils/random.js';
-import { expect } from '../../../test-helper.js';
+import { random } from '../../../../../src/shared/infrastructure/utils/random.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Utils | Random', function () {
   describe('#weightedRandoms', function () {
